SGE workaround
SGE overwrites the TMPDIR environment variable to a folder inside
the local file system, not NFS where the pipestance is running.
Thus, os.rename will crash as it only works in one filesystem.

Since we want to continue to use the atomicity of os.rename,
we will write the tmp journal file to the journal directory
but with the suffix ".tmp". Mario runtime will ignore and
not delete all these files.