Add a --notify option to marsoc and mrp.
The --notify options gives a single shell script that marsoc will run
when a pipestance ends for any reason (success or failure).  The script
will be run with the as:

<path-to-script> <pipestance-path> <success|failed> <error-message-pathr>

Any failures in running the script are ignored.

This also includes a prototype script "marsoc_onexit_handler.sh" that is
intended to be the argument to the notify flag in the marsoc environment.
The script runs the ligo_upload command with the path to the newly
terminated pipestance.

This also causes ligo_upload to bhe built when running "make" for marsoc.

This is a combined (but squashed) work from David & Haynes.