Fix serious bug introduced by --onfinish
The implementation for --onfinish created file-like objects that refered to
stdout and stderr manually to instantiate the subprocess.  When those objects
were GC'ed, go also closed the corresponding files. This caused stdout/stderr to
be inappropriately re-used and led to all kinds of crazyness if it raced with
another function that opened many file descriptors.

Test plan:
Create a go program with two go-routines: one that calls a script every 70ms and
another that crawls a big directory tree every 100 ms. Note that after a few
iterations the program crashes due to EBADF errors.  Adjust make this fix in the
goroutine that calls the script and note that the program no longer crashes.