Check that queued jobs are actually queued.
Allow command-line args to be set for qsub.

Set the "terse" argument for qsub, which causes its standard output to
be just the job id.

Record the job id from the qsub command.  Assume the arguments are set
appropriately on the command so output will be the jobid, and store it
in _jobid.

Before checking heartbeats, check every metadata which claims to be in
the queued state and has a jobid recorded.  If it is queued, run a script to
check which jobs are actually queued.  If the job is not queued, trick
the metadata into thinking the job is running for purposes of heartbeat.
This allows jobs to heartbeat-fail if either the jobs was errored-out by
the queue master, or if it ran but failed before it was able to write a
heartbeat, while still giving the job a full heartbeat interval after
leaving the queue to write its real heartbeat.