Fix grammar build for go 1.8
First of all, go yacc was moved out of the main go distribution and into
the go tools repo, so now we have to build it ourselves.  Second of all,
we were using some internal lexer variables which were changed around in
more recent versions of goyacc.  I've hacked around that issue, and the
tests still pass, so I'm declaring victory.

Also, setting travis to run "make all" so we don't get code rot like
this as often on the non-mrp things.