Fix two bugs in the formatter.
* Comments before `call` statements were getting eaten if the file only
contained calls.  Because mrf comments out include lines before
processing, this meant mrf was removing all includes from files which
contained only calls.  This is far from a complete fix for comment or
whitespace handling in mrf, but it helps.

* JSON formatted input args were not being formatted with indenting.

This change also applies mrf to all of the mro files in the tree,
although manually fixing up whitespace at the beginning of files was
still required.

Added a bunch of comments on test pipeline.  mrf does not format these
correctly yet.