Improve mrp status logging for long-running stages.
Adds an update_progress method to the python adapter, which allows
stages to push progress updates to the user.  The user will only see
these updates if mrp finds them before the stage completes or a
different progress update is written.

If any kind of update is written for a job (including log, heartbeat,
etc) and nothing has been written by that stage in at least 5 minutes,
mrp logs the current stage state.