Export more methods from core.
These are in support of the upcoming go-based job wrapper/monitor.

The main new exports are:
* HandleSignal (for registering as a signal handler)
* Metadata.MakePath
* Metadata.Write*
* Metadata.Read

Also add a new profile mode, "perf", to the ProfileMode constants (not
supported in the adapter yet, however).

Add a version of LogTee which takes an already open file.

Add a Metadata constructor which allows one to set a "run type" which
implies a prefix on the journal file names, and an UpdateJournal method
which can be used to write to the journal, and setting a files path
besides the default one (which is used for chunks of stages which don't
have splits or joins).

Strongly type stage code language.