Add a monitor process for martian jobs.
The monitor process does a few things:
1.  Starts the given subprocess which does the actual stage work, and
which usually also has to know how to interpret jobinfo and so on.
2.  Handles attaching perf record or pyflame profilers if appropriate.
3.  Handles the kneecapper and heartbeats.
4.  Gives the child process a pipe to write _errors or _asserts to.
5.  Handles some of the jobinfo rewriting, logging at the start and end
of the job, writing _complete if the job succeeded, and writing _errors
if it fails without writing an error of its own.

This also removes monitoring tasks from python adapter, and executes it
through the monitor.  This is a nearly complete rewrite of the python
code which, in conjunction with lots more comments, makes it pylint-clean.

In addition, this brings in a bunch of fixes to the adapter and related
stuff:

* Where appropriate, write metadata "atomically".  Specifically for the
jobinfo and progress files, there is a chance of a read race or corruption
due to an interupted write.

* In the python adapter make sure the journal is ALWAYS
updated for progress updates.

* In the python wrapper for subprocess.check_call and Popen, set PDEATHSIG.

This also provides a method for stage code to pass for a preexec_fn if
they're using the raw subprocess methods.

* Add a readme to adapeters.

This serves to document the API between the monitor and adapter process.
It also turns out to also be probably the best documentation we have in the
repo on how to actually write a stage.