Use a negative memory request amount as a sentinel.
Currently, stages which ask for more memory than was allowed locally
(with --localmem) are trimmed down to the local memory limit amount.
This change instead has them ask for a negative amount, which is
interpreted as a request for the larger of the absolute value of that
amount or the current local limit (which may be less than the maximum
local limit if less memory than expected is free).  If that amount is
larger than the maximum local limit, the stage fails.

Additionally, don't count mrp's memory usage against the
reservation/limit.