Fix mrf handling of comments and includes. (#18)
MRF was sometimes eating include lines and constantly attaching comments
to the wrong thing.  This should fix that.  It does so by treating
comments as a more first-class concept.  At the same time it gets rid of
that hacky thing with the lambda.

Also, tweak the argument alignment logic.

If one ID or help string is very long, don't let it blow out the line
length for every line in the list.  It's ok to have the one 80-character
help string out of alignment with everything else.