Run split/chunk/join in uniquified directories.
If a job gets accidentally double-started, it will do so in a new
directory.  The older job will be pointed at a directory which is
deleted and will fail rather than corrupting the state of the
still-running job.

This also includes updating the zip file utility to zip symlinks and
follow them when peaking inside a zip file.