Don't overwrite _errors with local job exit code.
If a local job fails with a non-zero exit code, the job manager writes
an _errors file stating that fact.  This change prevents it from doing
that if the file already exists, since that can potentially overwrite
useful debugging information.

In addition, add the messages mrjob produces when it catches a signal to
retry.json.  They are not the same as the python signal handler error messages.