Fix heartbeat for modes without queue check.
checkHeartbeat was changed to compare the last heartbeat time to the
last time that metadata was retrieved.  This should have been more
reliable in case some operation in between the scan of the journal and
the heartbeat check took a very long time (for example if nfs stalls for
an hour).  However, the last refresh time was only being set when the
job manager supported queue checking, since that was originally the only
thing that used that timestamp.  This resulted in the lastRefresh
timestamp being stuck at zero, which would never allow heartbeats to
time out.

This was not previously observed because heartbeat failures are uncommon
and nearly all of our internal testing happens with --jobmode=sge, which
supports the queue check.