Fix a possible deadlock during signal handling.
node.runJob uses a critical section to ensure that _queued_locally is
written 'atomically' with _jobinfo, to ensure restart works correctly.
It was not, however, releasing this critical section before calling
jobManager.execJob.  jobManager also uses a critical section to ensure
that _queued_locally is removed once the job is actually started.  On
some code paths, this is syncronous with the call from node.runJob.
Since a critical section can't be acquired while the signal handler is
waiting for existing critical sections to release, this results in a
deadlock where jobManager.execJob blocks indefinitely waiting to acquire a
critical section, and the signal handler can't exit because node.runJob
wouldn't release its critical section until execJob returned, if the
signal was received during that short window.