Dedupe direct prenodes.
Direct prenodes are used only for the UI.  By deduplicating them during
construction, we clean up the UI by avoiding having a separate arrow for
each output argument from one stage that is an input for another stage.
It also saves some memory.  Finally, we detect if something strange
happend and warn about it in the log rather than just forging on
blindly.

While we're at it, strongly-type more stuff in NodeInfo.