Allow binding from array elements.
* Enforce type match for all elements of arrays.

Previously, it was only enforced that the first element of the array was
of the correct type.

This is a compile-time check.  MRC passes for longranger, cellranger,
and cellranger-dna.

* Add constants for the various Kind types for Expression objects.

* Allow binding logic to recurse into arrays.

This allows bindings such as for example

call FOO (
    argv = [
        STAGE1.out,
        STAGE2.out,
    ],
)

* Allow trailing comma in array or map expressions.

This allows python-style lists, such as
[
   1,
   2,
   3,
],
to be used.

To maintain compatibility with older mrp versions, this final comma is
omitted by mrf, for now.  We'll add it back later, maybe, once we're
confident everyone is on an up to date mrp.