Skip waiting on split/join for non-split stages.
If the stage does not split, rather than setting the split's _complete and then
waiting for the next step to start the chunk, run the chunk immediately.
Likewise, when the chunk completes mark the stage complete immediately
instead of waiting for the next step to mark the join complete, and then
the step after that to mark the stage complete.  In total this saves 9
seconds on non-split stages, and also eliminates the (split_complete)
and (join_complete) messages from the log (which were confusing anyway,
for non-splitting stages).

For real pipelines, 9 seconds here and there is meaningless, but it sure
makes the integration tests go faster.