Only resolve stage bindings once per step.
Rather than re-evaluating the bindings for a stage for every chunk,
evaluate them once and share the resolved object between chunks.  For
stages with many chunks (for example some in Cellranger VDJ) this can
represent not only a substantial speedup in mrp performance, but also a
reduction in memory usage as the garbage collector is under less stress.