Add global enforcement level directive.
Hook into mrp and mrc.

Specifying --strict=error will cause pipelines to fail if they
violate newer type checks.  In alarm mode it will use the
existing alarm mechanism to report the error to the user.
In log mode it will only print to the log, and in disable mode
it will ignore the error.