Check for cyclic dependencies in compiler and mrf.
Apply a topological sort to pipeline calls.  It's a stable sort, meaning
calls will only be reordered in the case where a call's input is bound
to the output of a call that appears later in the source for that stage.

In many cases, report multiple errors from a single compilation run.

Add a bunch of parser unit tests.  Total 78.8% coverage.