Add syntax for specifying static resources in mro
For example,

    stage STAGENAME(
        in  int input1,
        out int output1,
    ) using (
        threads = 4,
    )

will reserve 4 threads for the main method of STAGENAME.

In addition, the 'using' keyword is no longer required for 'split',
allowing e.g.

    stage SPLITSTAGE(
        in  int[] input1,
        out int output1,
    ) split (
        in  int input_element,
        out int output_element,
    ) using (
        threads = 4,
        mem_gb = 1,
    )

which will reserve 4 threads and 1gb of memory for SPLITSTAGE's split phase,
as well as it's main and join if the split does not override those.