Introduce a new syntax for call modifiers.
This syntax allows us to bind call modifiers like we do inputs.  Right
now this is not supported for any of the existing call modifier types.
Preflight cannot be bound because it can't have dependencies.  Local is
not currently bindable because doing so would have complex interactions
with restart logic, and the use case doesn't seem to exist.  Volatile
also seems potentially problematic, and the imagined use cases for
making it bindable can be covered by overrides.