Throttle local jobs based on process ulimit.
Use a very rough estimate of process count per job to limit the number
of spawned jobs to stay within the ulimit.  With the current estimation
constants, which assume 15 processes per job plus one per reserved thread,
with the average ulimit of 1024, that allows job parallelism of 61 jobs in
local mode if nothing else is running.  One would hope that on machines with
64 more cores the ulimit would be set higher, but if not then this is better
than crashing at least.

Mrp will also periodically query the number of actual user processes and
the number being used by its process tree, just like it does for user
memory, in case some jobs use more than that rough estimate worth of
threads or in case the user is running other jobs on the machine.

If the number of processes available through ulimit is less than the
estimated minimum requrired to run with maximum parallelism given the
--localcores, a warning message is printed.