Fix inefficiency and race condition in maxjobs.
Previously, in cluster mode, for each running job there was a loop to
poll for job completion once per second.  This had two issues.  First,
it was inefficient, spawning O(maxjobs) goroutines which woke up once
per second.  Second, if a chunk failed and then, before the polling loop
woke up and discovered this, was reset and re-queued, the failed job
would never be considered completed and that slot in the semaphore would
be lost forever.

With this change, if the same metadata object tries to acquire multiple
times, subsequent acquires do not count against the limit.  This fixes
the race condition.

Second, rather than having a polling loop for each running job, a single
polling loop happens per top-level step.  Additionally, stages now
"push" notification of job completion to the semaphore.  These two mechanisms
are redundant, but that's ok since the polling is cheap.

Additionally, fix an oversight where mrs was not refreshing resources
periodically like mrp was.