Sync relevent files before notifying completion.
Specifically, sync the _outs file, the files directory entry, and the
metadata directory entry, in that order.  This may reduce the number
of cases where violations of total store ordering result in jobs running
on different servers seeing out of date output files, at least in the
most important cases.