"Partial" VDR on nodes depended on by top-level.
Rather than completely disabling VDR for nodes whos outputs are depended
on by the top-level pipeline, just turn off "volatile" for those stages.
Non-volatile stages do not have all of thier files deleted, but do have
their chunks /files deleted.  The behavior in Martian 2.3 was for VDR to
always be invoked on such stages - this is essentially a partial revert
of commit e52301680a34295f3b9fdfa0b6a17627554e6d7e.

On the other hand, don't consider completeness of the top-level pipeline
when deciding whether this non-volatile version of VDR is allowed to
run.  This allows VDR of the chunk files to happen earlier when in
"rolling" mode.