Track the IO bytes usage for stages.
This can be read from /proc/[pid]/io when crawling the process tree for
memory usage.  For jobs which last long enough to collect these stats,
the total as well as peak rate and rate standard deviation is collected.

In the final _perf file, the time-weighted standard deviation is
reported for individual jobs, but the standard deviation of the mean
rate across all jobs in the aggregation is reported.