Do not attempt to re-encode non-unicode strings.
Pass them to the file writer directly. Add a few new test cases (which fail on master).

If you pass a unicode to a python writer object, it attempts to encode the object as
ascii, resulting in an error if there are non-ascii characters in the message.

If we have a python str that's all ascii, or or a python unicode, then .encode() converts
it to a str which the writer can handle.

However, calling encode() on a str causes it to first try to decode() from ascii to unicode.
If the original string was not ascii, this fails. We could decode first, if we were sure we
always knew the encoding, but it's safer (in terms of not crashing) to just let the bytes
pass through unmolested.