Reorganize repository to be go-getable.
This means moving everything out of the src subdirectory to the
top-level.  The src directory now contains some symlinks to allow the
Makefile to build with the repository root as the GOPATH, but it will
now also be possible to build with the repository checked out at
$GOPATH/github.com/martian-lang/martian, either for the binaries or for
other tools using that gopath which want to import martian as a library.