Use template vars to speed the web UI a bit.
Use the template vars already provided by martian instead of waiting for
the json query to go through and then populating via javascript/angular.
This also reduces the number of things angular needs to hook up watches
for, which is nice.  The speedup is very slight overall, but it does
mean things like the pipestance ID render immediately rather than coming
in with the refresh when the graph renders.