Make the MroCache process multithreaded.
This has the potential danger that if a callable is defined in more than
one file, it is no longer as deterministic in terms of which one
actually gets cached.  However, since filesystem order is not generally
deterministic anyway, this was already a ticking time bomb, which we now
at least warn abount.

This also makes the MroCache object thread safe, which will be helpful
if we every want to start lazy-loading, for example.