Fix writing of invocation files.
This was broken when we stopped caching mro.  However, it's simple
enough to cache the callable for a stage.  It's also more correct, since
it is resilient to a stage being defined in more than one file.

If a stage is defined in more than one file, that's an error if both
files are in the transitive set of includes from the top-level call, but
if they're not then before we could end up silently getting the wrong
stage definition, which is confusing to debug.