Implement partial VDR, add per-node temp directory
This gets us higher resolution accounting for storage events, and allows
us to immediately clean up temp directories.  This also lays the
groundwork for strict-mode VDR.

In addition, even without strict-mode VDR, we can a full VDR immediately
if none of the depended-on output arguments of a stage have any strings
containing "/" in them, since without those, downstream stages have no
way to find the stage's files directory in order to implicitly depend on
it.