Wait for web requests to complete on shutdown.
Also, add error logging for the web server, and set various sensible
timeouts.

Also close a few cases where mrp could shut down (on purpose) without
waiting for critical sections to complete.

Also add a boolean argument to util.Suicide() to allow it to be used to
shut down the process with a successful exit code while still waiting
for critical sections and handlers to complete.