Add --includes option to mrf.
With --includes handling turned on, mrf will:
1.  Make sure that the file defining any stage or pipeline referenced by
the file being formatted is included directly by that file (if it was
transitively included).
2.  Remove any includes which do not define a stage or pipeline
referenced by the including file.
3.  Search the mropath for files defining stages or pipelines which were
not defined in the transitive closure of the referring file's includes.
4.  Add filetype definitions for any file types directly used in file
which are not already included in that file or its direct includes.

Step 1 is required in order to make step 2 safe. Otherwise file 1 could
include file 2, which defined stage A which is referenced by a pipeline in
file 3. If file 3 only includes file 1, then if file 1 drops the include
of file 2, now file 3 broken. Step 4 is required for similar reasons,
but because filetype redefinition is permitted, we don't want to pull in
extra files just for their filetype definitions. We can also usually assume
that when a pipeline uses a filetype, it's because on of its stages does, too,
so the direct inclusion requirement need only apply for types used in stage
args.

This also makes `mrf --rewrite` stop dumping to stdout.