Dynamically list top-level metadata files.
List the files which exist, and only the files which exist.  Also list
files in "extras", if present.  This involves adding two new API
endpoints, "/api/list-metadata-top" to get the list of top-level and
extras files, and "/extras/" to get the content of the files in extras.

Additionally, switch to useing http.FileServer for top-level and extras
files instead of going through the runtime object.  Top-level metadata
files aren't zipped, so we don't need to worry about that, and this way
has several efficiency advantages particularly when retrying or refreshing
a top-level metadata file.

Thirdly, update node dependencies for the web build, including updating
to gulp 4 to silence some security warnings from npm.