Lazily deserialize args/outs.
Most of the time, when we read args/outs, we're just copying some or all
of the keys to another json file.  Therefore, rather than going to the
effort to fully decode it and then re-encode it, only decode one level
down into the map.  We still have to fully decode when validating
outputs, but firstly that is only done once (as opposed to every time we
need to build a dependent stage's args), and second that can be done one
key at a time and then discarded, saving memory.  Meanwhile we get to
skip the effort of re-serializing when we write.

Along the way fix a few other minor inefficiencies.