Run monitor's memory check loop much more often.
Running every 5 seconds instead of every 30 means more likely to catch
the high water mark for total RSS.

Also, carefully tune perf_unix.go to eliminate many cases (other than
error messages) where temporary variables escape to the heap.  The
number of allocations in the benchmark (which was added in this change)
is reduced by 40%, and the speed is increased by 10-20%.

Optimize a few other cases while we're at it.