Semantically compare ASTs on reattach.
When reattaching to a pipestance, do a deep semantic comparison of the
old _mrosource with the one which would be generated now.  The
comparison allows some things to change, such as VDR mode or resource
requests, but does not allow changes which would alter the pipestance
directory structure, such as adding or removing calls to a pipeline,
changing a call alias, or changing whether a stage splits.  It will also
reject changes which might invalidate type checking, such as changes to
the input or output parameters of stages or pipelines.  Changes to
hard-coded argument values will also be rejected.

In addition, correct the creation of @include directives in _invocations.
Rather than adding the complete transitive closure of includes for the
top-level invocation, only add the include which defines the stage or
subpipeline being called.  This fixes issues with invalid _invocation
files being generated when a file was included by multiple files in the
transitive include closure of the top-level file.  It also makes it
easier to figure out from the invocation file what is actually being
called.