Remove mrs.  Teach mrp to run pipelines.
Instead of maintaining a separate (and less functional) program for
running individual stages, allow mrp to wrap a stage and run it as a
pipeline.

In addition to simplifying code maintenance, this also means that now
mrs gets partial restart, a web UI, and so on and so forth.

mrs still exists as a symlink to mrp.