Allow preflights to be disabled.
Add some tests to make sure that preflight stages cannot depend on other
stages in the pipeline (which would be a circular dependency), including
for disablement.

Also add compile check for a pipeline which calls itself, and unit tests
to exercise that.  If a pipeline did call itself, the infinite recursion
would result in a crash when constructing the dependency graph.  Maybe
eventually we can add support for more controlled pipeline recursion but
for now, nope.