Make stages more robust against journaling failure.
Martian determines state for a metadata object based on the presence or
absence of certain files.  Rather than polling the directory of every
active chunk looking for new files, however, it just polls the journal
directory, where the job should be dropping a notification of new files.
However, various things can prevent the journal from being written even
if the file was written successfully.

In the rare event that a stage is about to fail due to heartbeat error,
or if the job manager says the job has finished, directly poll the job's
metadata directory looking for a state change first.  If the job has
completed (or failed), ignore the fact that the journal entry was
missed.

Similarly, when we see the split phase has completed but haven't seen a
journal update for the _stage_defs file, immediately poll for it rather
than waiting until it would heartbeat out before trying to poll it.

Because these cases are relatively rare, the impact of the extra filesystem
operations should be minor, but might help with several edge cases
people seem to be getting into.