Account for virtual address space size.
First, rather than decomposing JobResources into its constituent parts
in several places, pass around the whole object instead.

Allow stages to specify separate mem_gb and vmem_gb amounts.  Just like
mem_gb, this can be set by:
 * setting "__vmem_gb" in stage_defs returned by the split.
 * setting `using ( vmem_gb = 123 )` in the mro defining the stage.
 * in overrides.json, using e.g. "split.vmem".

If not specified, the default vmem_gb is mem_gb plus a constant value
configured in config.json (currently 3).  If it is specified, but less
than the rss value, it will be overridden to be equal to the rss value.

In local mode, vmem is unlimited unless a vmem ulimit is detected, or
the --localvmem flag is set.  If --localvmem is greater than the ulimit
minus the usage for mrp itself, the ulimit will be used instead.  A
warning is issued if the vmem limit is insufficient for a default-sized
job.

In cluster mode, new __MRO_VMEM_GB__ and related template parameters are
available.  If the config.json for each cluster job mode, the setting
    "mem_is_vmem": true
then __MRO_MEM_GB__ and related parameters will use vmem values as well.
This is set to true for SGE and derivatives.  Vmem does NOT affect thread
counts when using --mempercore, since that would not make any sense.