Optimizations to computation of files to args map.
The computation of which arguments bind each file for VDR is
O(number of files referenced in output args * number of files produced)
because of the need to check subdirectories.  This optimizes (to O(1)) the
fairly common case where most of the files are directly referenced in the
outs, as well as making some minor improvements to the coefficient in
other cases and improving memory allocation performance.

This does not help (much) in the case where the outs specify a directory
containing a large number of files, nor the case where a large number of
output files are not referenced in the outs (but the number referenced
in the outs is also still large).  That case could be solved by building
a prefix tree, but it is sufficiently uncommon to not justify adding
that complexity, at least for now.

This speeds up the benchmark introduced in the last commit, which has
200 files referenced by one of the output parameters, by a factor of 7.
The larger the that number of files, of course, the larger the speedup.