Periodically verify that the log file still exists.
If the pipestance directory is deleted, shut down. Also shut down if the
log file still exists but isn't the same file as the one mrp has open,
as that can happen if the pipestance was deleted but then another mrp
instance started up.  This serves as a type of backstop if the _lock
file gets deleted when it shouldn't be.