Detect and report memory cgroup limits.
Use them for the default RSS limit.  If `--localmem` is set, that takes
precedence, but `mrp` will warn the user if the specified value is
greater than the detected cgroup limit.

This is particularly useful for users who queue their `mrp` job to SLURM
clusters with `bsub`, which sets the `hierarchical_memory_limit` in
`memory.stat` for the process cgroup.  This also allows better reporting
and detection of actual free memory in a cgroup-limited context.

In order to make the `--localmem` escape valve more effective, in case
there are odd configurations where cgroup limits are detected incorrectly,
the cgroup memory RSS is not used for purposes of throttling (actual
system free memory is uesd that way).  However, mrp will warn if the
cgroup limit minus the cgroup rss is less than the total free system memory
and less than what it expects to have available.