Prepend rather than append stage code python path.
This hopefully prevents accidentally importing the wrong module in the
event that there is a module somewhere in the PYTHONPATH with the same
name as the base name of the stage code module.

Actually, overwrite the current working directory default first path
with the stage code path, because that doesn't make sense in context
either.