Add struct and typed map support.
This involved a complete rewrite of type system.  Arrays are now types
themselves, and "type names" are now a struct which can represent the
array or map properties of the type.  Types now know how to validate
expressions and perform type conversions.

syntax.ValExp is no longer a struct type.  Instead it is an interface,
implemented by expression types for each kind.  This makes the handling
of for example array expressions which contain RefExp elements much
cleaner.

Refactor a bunch of existing code, splitting it into multiple files.

Significantly improve test coverage, and fix a bunch of linter errors.
Also clean up some tests, including removing extraneous warnings output.

Add call resolution methods to syntax package.

These methods allow one to generate the call graph statically, without
needing to instantiate a pipestance.  Inputs for stages and pipelines,
as well as outputs of pipelines, are resolved to either a literal
expression or the output of a stage (not another pipeline).

Add --graph option for mrc, which outputs the resolved call graph in json form.

Fix a bug in check that all inputs to a pipeline are used.

Fix Type.IsFile semantics to be more useful. Specifically, replace the boolean
value with one of 4 values, either not a file, not a formally defined file but
maybe containing paths (like a string) or a type which contains more
than one file (such as a directory or struct).

Remove non-lazy ArgumentMap type.  Add a new MarshalerMap type, which
works like a LazyArgumentMap but allows lazily serializing as well as
deserializing, for key types (like syntax.Exp) which support it.

Remove obsolete ToInterface method from Exp.  Instead, expressions now
just implement MarshalJSON().

Add a BuildCallSource method to core.InvocationData.  This essentially
exposes the functionality that `mrg` has.