Rework mro include path resolution.
If a given @include directive could refer to multiple files in the mro
path, raise an error instead of just taking the first one.  Otherwise
pipeline authors can get some difficult-to-debug behavior when the file
they included wasn't the one they thought they were including.

Add an optional field to InvocationData, named `mro_file` in the json form,
to specify the file from which the given callable is included, rather than
forcing tools like `mrg` to search through the `MROPATH` until they find a
file which defines the requested pipeline.  Populate that field when
generating an InvocationData from an ast.