Rewrite the top-level _outs with updated paths.
As part of this, split the logic for moving files apart from the logic
for printing outputs for the user.  This will result is a bit of
inefficiency because some things get unmarshalled more than once.  On
the other hand, it should make maintenance easier.

After moving the output files into outs/, update _outs json file to have
those paths.  Otherwise, if the rest of the pipestance is blown away and
only the top-level metadata and the outs folder is kept, the symlinks
which the _outs json was pointing to will no longer be valid.

Enhance the tests for post_process_test to check that this rewrite is
happening.