Correctly issue warnings for strict=alarm or log.
Setting strict=alarm or log was supposed to issue warnings for
strict-mode violations, but was instead issueing errors for bad chunk
definitions or outputs.  This makes migration difficult, as one can't
run a pipeline to completion and then collect all of the errors to fix
at once, since it'll keep on failing part-way through.  This restores
the intended behavior of allowing the pipestance to continue in these
cases.