Fix some issues with invalid outputs.
First of all, correct an inconsistency between toleration of non-string
outputs for user-defined file types when validating outputs vs when
resolving inputs.

Second of all, if errors are encountered when resolving inputs, provide
a null value and continue rather than providing an incomplete map.

If an error is encountered when resolving inputs, rather than just
ignoring it (and attempting to run with a possibly-incomplete argument
binding structure), report it according to the strict enforcement level.

Also improve the traceability of error messages in a few cases.