Add support for map call and split inputs.
This feature is intended to replace sweep, which will be dropped in a
subsequent commit.

When calling a stage or sub-pipeline, instead of just `call STAGE(...)`
one can now `map call STAGE(...)`.  One or more input parameters to the
stage or sub-pipeline must then specified as being split, e.g.

```
map call STAGE(
    foo = split GENERATOR.array,
)
```

That input must be an array or typed map of the expected input type.
The stage or sub-pipeline is then called once for each element in the
array or map.  In the case of sub-pipelines, only stages which depend on
the split inputs are re-run.  The result of this mapped call is then an
array or typed map (with the same keys) of struct values of the type
normally produced by the stage or pipeline.  If multiple inputs are
split, they must either all be arrays with the same length, or they must
all be typed maps with the same keys.

For an example use case, imagine,
```
pipeline ANALYZE(
    in Sample sample,
    ...
)

pipeline AGGREGATE(
    in map<ANALYZE> analyses,
)

pipeline META_ANALYSIS(
    in map<Sample> samples,
    ...
)
{
    map call ANALYZE(
        sample = split samples,
        ...
    )

    call AGGREGATE(
        analyses = ANALYZE,
    )
...
}
```

One could also rewrite a traditional stage with a split/main/join
as something like
```
call STAGE_SPLIT(...)

map call STAGE_MAIN(
    input1 = split STAGE_SPLIT.chunk_defs,
)

call STAGE_JOIN(
    chunk_outs = STAGE_MAIN,
)
```
While this might not be terribly interesting, it becomes more
interesting if one would wishes to "chain" stages together with similar
chunking, without a pointless join/split in between.

Without mapping any stages, this feature also allows one to reshape a set
of outputs, e.g.
```
pipeline IDENTITY(
    in  file foo,
    in  file bar,
    in  file baz,
    out file foo,
    out file bar,
    out file baz,
)
{
    return (
        foo = self.foo,
        bar = self.bar,
        baz = self.baz,
    )
}

pipeline RESHAPE(
    in file[] foo,
    in file[] bar,
    in file[] baz,
    out IDENTITY[] values,
)
{
    map call IDENTITY(
        foo = split self.foo,
        bar = split self.bar,
        baz = split self.baz,
    )

    return (
        values = IDENTITY,
    )
}
```
to convert three arrays into an array of structs.