Fix subforks.
When creating the forks of a node with Node.buildForks(), it would
add forks corresponding to Node.subnodes.  That got broken during the
struct type changes.  Rather than bring that back, since it's redundant
information, just iterate through the subnodes and match forks as
needed.

This was breaking perf data reporting.