Update how stage outs file is pre-populated.
Do not use the broken backwards-compatibility behavior of pre-populating
a string for an array of file type objects on typed maps or structs.

Pre-populate arrays with empty arrays rather than nil, and pre-populate
typed maps with empty maps rather than nil.

For struct types, recursively pre-populate the maps with the proper
fields.