Fix an crash for pipelines with constant outputs.
These are more reasonable than you might think at first glance, as they
could be used to define constants or to populate structs with repetitive
information.  Whether that's a good idea or just a hack is open to
debate, but it can certainly be useful for declaring placeholders during
development.