Permit call-less pipelines.
Also permit pipeline inputs which are only bound to outputs, without a
call using them.

This feature is useful for creating a reusable "pipeline" as way to
reusably pack a set of inputs into a structure.

Also modify the mrc --graph output to not HTML-escape.  We're not
embedding in html most of the time and it makes type names for typed
maps annoying to read.