Support wildcard argument bindings.
These work for both call and return bindings.  The form of a wildcard
binding is to use a '*' for the parameter name, e.g.

    call (
        z = self.something,
        * = self.struct_input,
    )

This expands out to be equivilent to writing

    call (
        z = self.something,
        x = self.struct_input.x,
        y = self.struct_input.y,
    )

The right-hand side must be a reference to a struct (or array or typed map of
structs), and every required parameter which has a corresponding field in the
struct will be set to that value.  It is not permitted to use a wildcard on the
left-hand side if some of the arguments had previously been bound (in the
example above, if struct_input had a member named z).  The wildcard binding
must be the last binding in the binding list, and there can only be one
of them.