Don't locally redefine structs in mrf.
For user-defined file types defined in an mro file which doesn't
directly include the defining file, we make local definitions for the
file types (and possibly remove the @include for them if it's no longer
in use).  However we don't want to do this for struct types, since they
have structure beyond the name.