Python journal errors -> mostly just warnings.
The actual important journal entries are written by go code, and even
those aren't really critical.  Some error types are expected and ignored
completely, the rest are warnings in case they might shed light on later
problems.  The only problem that causes a failure is ENOENT, which is
what we'd get if the journal directory was deleted (meaning the
pipestance already completed or was deleted).