Allow direct inclustion of transitively requirements.
As a general rule, depending on something transitively is not a good
idea.  However, if a stage depends on a type that is included
transitively, it probably does so because the type is used by a stage
which it depends on directly, so it's ok to pick up the type
transitively that way.  However, we don't want mrf to remove the include
in that case.